1 实验源码/2 基于深度学习版 （Bi-LSTM+CRF）/requirements.txt
pytorch-crf
scikit-learn